pyproject.toml
[project]
name = "finalproject"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pillow>=12.1.0",
    "streamlit>=1.52.2",
    "ultralytics>=8.3.247",
]
